+++ routes/api/obras.php
@@ -5,6 +5,7 @@ use App\Http\Controllers\IncomeController;
 use App\Http\Controllers\ObraController;
 use App\Http\Controllers\ObraDailyLogController;
 use App\Http\Controllers\ObraDailyLogTagController;
+use App\Http\Controllers\ObraDocumentController;
 use App\Http\Controllers\ObraStageController;
 use App\Http\Controllers\ObraStageTaskController;
 use App\Http\Controllers\OutcomeController;
@@ -13,50 +14,59 @@ use Illuminate\Support\Facades\Route;
 
 // Obras endpoints
 Route::prefix('obras')->middleware('auth:sanctum')->controller(ObraController::class)->group(function () {
-  Route::get('/', 'index');
-  Route::get('/{id}', 'show');
-  Route::get('/{id}/contractors', 'contractors');
-  Route::post('/', 'store');
-  Route::post('/{id}', 'update');
-  Route::delete('/{id}', 'destroy');
-  Route::post('/{id}/documents', 'documents');
-  Route::post('/{id}/additionals', 'additionals');
+  Route::get('/', 'index')->middleware('permission:obras_list');
+  Route::get('/{id}', 'show')->middleware('permission:obras_display');
+  Route::post('/', 'store')->middleware('permission:obras_insert');
+  Route::post('/{id}', 'update')->middleware('permission:obras_update');
+  Route::delete('/{id}', 'destroy')->middleware('permission:obras_delete');
+  
+
+  Route::get('/{id}/contractors', 'contractors')->middleware('permission:obraContractors_list');
+  Route::post('/{id}/additionals', 'additionals')->middleware('permission:obraAdditional_insert');
 });
 
 // Incomes endpoints
-Route::prefix('incomes')->middleware('auth:sanctum')->controller(IncomeController::class)->group(function () {
-  Route::get('/', 'index');
-  Route::get('/export', 'exportList');
-  Route::get('/{id}', 'show');
-  Route::post('/', 'store');
-  Route::post('/{id}', 'update');
-  Route::delete('/{id}', 'destroy');
+Route::prefix('obras/{obraId}/incomes')->middleware('auth:sanctum')->controller(IncomeController::class)->group(function () {
+  Route::get('/', 'index')->middleware('permission:obraIncomes_list');
+  Route::get('/list_all', 'listAll')->middleware('permission:obraIncomes_list');
+  Route::get('/export', 'exportList')->middleware('permission:obraIncomes_export');
+  Route::get('/{incomeId}', 'show')->middleware('permission:obraIncomes_display');
+  Route::post('/', 'store')->middleware('permission:obraIncomes_insert');
+  Route::post('/{incomeId}', 'update')->middleware('permission:obraIncomes_update');
+  Route::delete('/{incomeId}', 'destroy')->middleware('permission:obraIncomes_delete');
 });
 
 // Outcomes endpoints
-Route::prefix('outcomes')->middleware('auth:sanctum')->controller(OutcomeController::class)->group(function () {
-  Route::get('/', 'index');
-  Route::get('/export', 'exportList');
-  Route::get('/{id}', 'show');
-  Route::post('/', 'store');
-  Route::post('/{id}', 'update');
-  Route::delete('/{id}', 'destroy');
+Route::prefix('obras/{obraId}/outcomes')->middleware('auth:sanctum')->controller(OutcomeController::class)->group(function () {
+  Route::get('/', 'index')->middleware('permission:obraOutcomes_list');
+  Route::get('/list_all', 'listAll')->middleware('permission:obraOutcomes_list');
+  Route::get('/export', 'exportList')->middleware('permission:obraOutcomes_export');
+  Route::get('/{outcomeId}', 'show')->middleware('permission:obraOutcomes_display');
+  Route::post('/', 'store')->middleware('permission:obraOutcomes_insert');
+  Route::post('/{outcomeId}', 'update')->middleware('permission:obraOutcomes_update');
+  Route::delete('/{outcomeId}', 'destroy')->middleware('permission:obraOutcomes_delete');
+});
+
+// Documents endpoints
+Route::prefix('obras/{obraId}/documents')->middleware('auth:sanctum')->controller(ObraDocumentController::class)->group(function () {
+  Route::get('/', 'index')->middleware('permission:obraDocuments_list');
+  Route::post('/', 'store')->middleware('permission:obraDocuments_insert');
 });
 
 // Additionals endpoints
 Route::prefix('additionals')->middleware('auth:sanctum')->controller(AdditionalController::class)->group(function () {
-  Route::get('/{id}', 'show');
-  Route::post('/{id}', 'update');
-  Route::delete('/{id}', 'destroy');
+  Route::get('/{id}', 'show')->middleware('permission:obraAdditional_display');
+  Route::post('/{id}', 'update')->middleware('permission:obraAdditional_update');
+  Route::delete('/{id}', 'destroy')->middleware('permission:obraAdditional_delete');
 });
 
 // ObraDailyLog endpoints
 Route::prefix('obras/{obraId}/daily_logs')->middleware('auth:sanctum')->controller(ObraDailyLogController::class)->group(function () {
-  Route::get('/', 'index');
+  Route::get('/', 'index')->middleware('permission:obrasDailyLogs_list');
   Route::get('/{dailyLogId}', 'show');
   Route::get('/{dailyLogId}/file', 'fileDownload');
-  Route::post('/', 'store');
-  Route::post('/{dailyLogId}', 'update');
+  Route::post('/', 'store')->middleware('permission:obrasDailyLogs_insert');
+  Route::post('/{dailyLogId}', 'update')->middleware('permission:obrasDailyLogs_update');
 });
 
 // ObraDailyLogTag endpoints
@@ -65,20 +75,21 @@ Route::prefix('obra_daily_log_tags')->middleware('auth:sanctum')->controller(Obr
   Route::get('/{id}', 'show');
 });
 
+// Tasks in stages endpoints
 Route::prefix('obras/{obraId}/stages/{stageId}/tasks')->middleware('auth:sanctum')->controller(ObraStageTaskController::class)->group(function () {
-  Route::get('/', 'index');
-  Route::get('/{taskId}', 'show');
-  Route::post('/', 'store');
-  Route::post('/{taskId}', 'update');
+  Route::get('/', 'index')->middleware('permission:obraStageTasks_list');
+  Route::get('/{taskId}', 'show')->middleware('permission:obraStageTasks_display');
+  Route::post('/', 'store')->middleware('permission:obraStageTasks_insert');
+  Route::post('/{taskId}', 'update')->middleware('permission:obraStageTasks_update');
   Route::post('/{taskId}/completed', 'checkCompleted');
-  Route::delete('/{id}', 'destroy');
+  Route::delete('/{id}', 'destroy')->middleware('permission:obraStageTasks_delete');
 });
 
 // Stages endpoints
 Route::prefix('obras/{obraId}/stages')->middleware('auth:sanctum')->controller(ObraStageController::class)->group(function () {
-  Route::get('/', 'index');
-  Route::get('/{stageId}', 'show');
-  Route::post('/', 'store');
-  Route::post('/{stageId}', 'update');
-  Route::delete('/{id}', 'destroy');
+  Route::get('/', 'index')->middleware('permission:obraStages_list');
+  Route::get('/{stageId}', 'show')->middleware('permission:obraStages_display');
+  Route::post('/', 'store')->middleware('permission:obraStages_insert');
+  Route::post('/{stageId}', 'update')->middleware('permission:obraStages_update');
+  Route::delete('/{id}', 'destroy')->middleware('permission:obraStages_delete');
 });
